requirement.txt
pip install django 
 django-admin startproject tutorial
 python manage.py runserver 